Show alert dialog when turn off notifications clicked
