+++ notifications/src/main/kotlin/uk/gov/govuk/notifications/ui/NotificationsConsentScreen.kt
@@ -14,7 +14,6 @@ import uk.gov.govuk.notifications.NotificationsConsentViewModel
 import uk.gov.govuk.notifications.NotificationsUiState
 import uk.gov.govuk.notifications.NotificationsViewModel
 import uk.gov.govuk.notifications.R
-import uk.gov.govuk.notifications.openDeviceSettings
 
 @OptIn(ExperimentalPermissionsApi::class)
 @Composable
@@ -57,7 +56,10 @@ internal fun NotificationsConsentRoute(
                             secondaryText = secondaryText,
                             onSecondary = {
                                 notificationsViewModel.onTurnOffNotificationsClick(secondaryText)
-                                openDeviceSettings(context)
+                                showNotificationsAlert(
+                                    context,
+                                    onCancelButtonClick = { notificationsViewModel.onCancelButtonClick(it) },
+                                    onContinueButtonClick = { notificationsViewModel.onContinueButtonClick(it) })
                             }
                         )
                     }
