+++ notifications/src/main/kotlin/uk/gov/govuk/notifications/ui/NotificationsPermissionScreen.kt
@@ -1,7 +1,5 @@
 package uk.gov.govuk.notifications.ui
 
-import android.app.AlertDialog
-import android.content.Context
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
@@ -9,7 +7,6 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
-import androidx.core.app.NotificationManagerCompat
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import uk.gov.govuk.design.ui.component.ChildPageHeader
@@ -18,7 +15,6 @@ import uk.gov.govuk.notifications.NotificationsUiState
 import uk.gov.govuk.notifications.NotificationsPermissionViewModel
 import uk.gov.govuk.notifications.NotificationsViewModel
 import uk.gov.govuk.notifications.R
-import uk.gov.govuk.notifications.openDeviceSettings
 
 @OptIn(ExperimentalPermissionsApi::class)
 @Composable
@@ -75,8 +71,8 @@ internal fun NotificationsPermissionRoute(
                 val context = LocalContext.current
                 showNotificationsAlert(
                     context,
-                    onCancelButtonClick = { viewModel.onCancelButtonClick(it) },
-                    onContinueButtonClick = { viewModel.onContinueButtonClick(it) })
+                    onCancelButtonClick = { notificationsViewModel.onCancelButtonClick(it) },
+                    onContinueButtonClick = { notificationsViewModel.onContinueButtonClick(it) })
                 viewModel.finish()
             }
 
@@ -85,35 +81,5 @@ internal fun NotificationsPermissionRoute(
                 notificationsPermissionCompleted()
             }
         }
-    }
-}
-
-private fun showNotificationsAlert(
-    context: Context,
-    onContinueButtonClick: (String) -> Unit,
-    onCancelButtonClick: (String) -> Unit
-) {
-    val isNotificationsOn = NotificationManagerCompat.from(context).areNotificationsEnabled()
-    val alertTitle =
-        if (isNotificationsOn) R.string.notifications_alert_title_off else R.string.notifications_alert_title_on
-    val alertMessage =
-        if (isNotificationsOn) R.string.notifications_alert_message_off else R.string.notifications_alert_message_on
-    val cancelButton = context.getString(R.string.cancel_button)
-    val continueButton = context.getString(R.string.continue_button)
-
-    AlertDialog.Builder(context).apply {
-        setTitle(context.getString(alertTitle))
-        setMessage(context.getString(alertMessage))
-        setNeutralButton(cancelButton) { dialog, _ ->
-            onCancelButtonClick(cancelButton)
-            dialog.dismiss()
-        }
-        setPositiveButton(continueButton) { dialog, _ ->
-            onContinueButtonClick(continueButton)
-            openDeviceSettings(context)
-            dialog.dismiss()
-        }
-    }.also { notificationsAlert ->
-        notificationsAlert.show()
     }
 }
